pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adbc-driver-postgresql>=1.8.0",
    "emoji>=2.14.1",
    "graphviz>=0.21",
    "jupyter>=1.1.1",
    "pandas>=2.3.2",
    "polars>=1.32.3",
    "praw>=7.8.1",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
    "textacy>=0.13.0",
]
